Update ssl_wae_gma_rvi_multi.mq5
While debugging I ran into a few errors and functionality problems. The first error was accessing an empty array. Fix: Call the function to fill the array. 

Functionally, I wasn't placing trades, so I had to add that code. Additionally, I changed the no trade days to less so that I could get a trade going. Next Functionality issue was it wasn't accessing the next currency due to off by 1 error so I set the <= operator in the loop. Still need to test. I am thinking I add volume next and then create the money management portion. Not sure if the money management portion should be inheritance/polymorphism or not. I'll ask the code mentor for suggestion on how to refactor that for efficiency.